+++ hardware/include/nb/BasicWorldConstants.h
@@ -1,7 +1,7 @@
 #ifndef BasicWorldConstants_h
 #define BasicWorldConstants_h
 
-static const float GRAVITY_mss = -9.8f; //in Z direction (up is positive)m/s^2
+static const double GRAVITY_mss = -9.8; //in Z direction (up is positive)m/s^2
 
 static const int DAYS_PER_YEAR = 365; //Except for leap years.... uh oh
