+++ hardware/include/nb/InverseKinematics.h
@@ -9,18 +9,18 @@ namespace Kinematics{
      */
     //Accuracy constants for dls
     //in mm, how close dls will get to the target
-    static const float UNBELIEVABLY_LOW_ERROR = 0.01f; //mm
-    static const float REALLY_LOW_ERROR = 0.1f; //mm
-    static const float ACCEPTABLE_ERROR = 0.5f; //mm
-    static const float COARSE_ERROR     = 1.0f; //mm
+    static const double UNBELIEVABLY_LOW_ERROR = 0.01; //mm
+    static const double REALLY_LOW_ERROR = 0.1; //mm
+    static const double ACCEPTABLE_ERROR = 0.5; //mm
+    static const double COARSE_ERROR     = 1.0; //mm
 
-    static const float dampFactor = 0.4f;
-    static const float maxDeltaTheta = 0.5f;
+    static const double dampFactor = 0.4;
+    static const double maxDeltaTheta = 0.5;
     static const int maxAnkleIterations = 60;
     static const int maxHeelIterations = 20;
 
 
-    static const float HYP_NOT_SET = -1000.0f;
+    static const double HYP_NOT_SET = -1000.0;
 
     enum IKOutcome {
         STUCK = 0,
@@ -29,19 +29,19 @@ namespace Kinematics{
 
     struct IKLegResult {
         IKOutcome outcome;
-        float angles[6];
+        double angles[6];
     };
 
     const IKLegResult simpleLegIK(const ChainID chainID,
                                   const NBMath::ufvector3 & legGoal,
-                                  float startAngles []);
+                                  double startAngles []);
 
     const IKLegResult legIK(const ChainID chainID,
                             const NBMath::ufvector3 &footGoal,
                             const NBMath::ufvector3 &footOrientation,
                             const NBMath::ufvector3 &bodyGoal,
                             const NBMath::ufvector3 &bodyOrientation,
-                            const float HYPAngle = HYP_NOT_SET);
+                            const double HYPAngle = HYP_NOT_SET);
     /**
      * Wrapper method for IK which finds leg angles given a location
      * for the leg, and angles for the body relative to the world
@@ -49,18 +49,18 @@ namespace Kinematics{
      */
     const IKLegResult angleXYIK(const ChainID chainID,
                                 const NBMath::ufvector3 & legGoal,
-                                const float bodyAngleX,
-                                const float bodyAngleY,
-                                const float HYPAngle);
+                                const double bodyAngleX,
+                                const double bodyAngleY,
+                                const double HYPAngle);
 
     const void clipChainAngles(const ChainID id,
-                               float angles[]);
-    const float getMinValue(const ChainID id, const int jointNumber);
-    const float getMaxValue(const ChainID id, const int jointNumber);
+                               double angles[]);
+    const double getMinValue(const ChainID id, const int jointNumber);
+    const double getMaxValue(const ChainID id, const int jointNumber);
     const NBMath::ufvector3 forwardKinematics(const ChainID id,
-                                              const float angles[]);
+                                              const double angles[]);
     const NBMath::ufmatrix3 buildJacobians(const ChainID chainID,
-                                              const float angles[]);
+                                              const double angles[]);
 
 
     // Both adjustment methods return whether the search was successful.
@@ -68,17 +68,17 @@ namespace Kinematics{
     // startAngles by reference.
     const bool adjustAnkle(const ChainID chainID,
                            const NBMath::ufvector3 &goal,
-                           float startAngles[],
-                           const float maxError);
+                           double startAngles[],
+                           const double maxError);
     const bool adjustHeel(const ChainID chainID,
                           const NBMath::ufvector3 &goal,
-                          float startAngles[],
-                          const float maxError);
+                          double startAngles[],
+                          const double maxError);
     const IKLegResult dls(const ChainID chainID,
                           const NBMath::ufvector3 &goal,
-                          const float startAngles[],
-                          const float maxError = ACCEPTABLE_ERROR,
-                          const float maxHeelError = UNBELIEVABLY_LOW_ERROR);
+                          const double startAngles[],
+                          const double maxError = ACCEPTABLE_ERROR,
+                          const double maxHeelError = UNBELIEVABLY_LOW_ERROR);
 
 
     const IKLegResult analyticLegIK(const ChainID chainID,
@@ -86,11 +86,11 @@ namespace Kinematics{
                                     const NBMath::ufvector3 &footOrientation,
                                     const NBMath::ufvector3 &bodyGoal,
                                     const NBMath::ufvector3 &bodyOrientation,
-                                    const float givenHYPAngle = HYP_NOT_SET);
+                                    const double givenHYPAngle = HYP_NOT_SET);
 
 
-    NBMath::ufmatrix4 rotationHYPRightInv(const float HYP);
-    NBMath::ufmatrix4 rotationHYPLeftInv(const float HYP);
+    NBMath::ufmatrix4 rotationHYPRightInv(const double HYP);
+    NBMath::ufmatrix4 rotationHYPLeftInv(const double HYP);
 
 };
 #endif
