+++ hardware/include/nb/WalkController.h
@@ -32,15 +32,15 @@ public:
     //WalkController(Sensors *s) : sensors(s) { }
     virtual ~WalkController() { };
 
-    virtual const float tick(const std::list<float>* zmp_ref,
-                             const float cur_zmp_ref,
-                             const float sensor_zmp) = 0;
+    virtual const double tick(const std::list<double>* zmp_ref,
+                             const double cur_zmp_ref,
+                             const double sensor_zmp) = 0;
 
-    virtual const float getPosition() const = 0;
+    virtual const double getPosition() const = 0;
 
-    virtual const float getZMP() const = 0;
+    virtual const double getZMP() const = 0;
 
-    virtual void initState(float x, float v, float p) = 0;
+    virtual void initState(double x, double v, double p) = 0;
 
 };
